requirements.txt
fluent-logger==0.4.4
PyYAML==3.12
